requirements.txt
fastapi
uvicorn
httpx
python-dotenv
eth-account
pynacl
requests
websockets
